scripts/starfield/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyvirtualcam>=0.11.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
httpx>=0.25.0
